Move table rendering to display mod
